qidian/qidian_ticket/requirements.txt
pyyaml
numpy
pyautogui
tqdm
imapclient
requests
Pillow
opencv-python
configargparse